Major fixes and cleanups to desktop.net package
- Major code modifications to cleanly close any operations. Calls to
send "CLOSE" messages where needed were added in InetAdapter, Connect,
StatusCheck and Finalize.
- Added a stray socket flag to said classes to prevent sockets from
staying open after a reset.
- Fixed unused message constants in OutThread and InThread
- MultiplayerSystem now sends "CLOSE" back when receiving a "CLOSE".
This was to reflect a subtle problem with the protocol whereas it was
not specified if "CLOSE" should destroy both ends of the communication
or only one. From now on we shall assume "CLOSE" means to close my end
of the channel, so if I receive a CLOSE from a friend I must close my
end as well.
- Fixed missing javadocs where needed.
- Added a TAG to MultiplayerSystem.